Move secureApi to shell
