+++ js/src/shell/secureApi.js
